+++ filament/backend/src/vulkan/VulkanDriver.h
@@ -49,10 +49,6 @@
 #include <utils/Allocator.h>
 #include <utils/compiler.h>
 
-#if defined(__ANDROID__)
-#include "AndroidNativeWindow.h"
-#endif
-
 namespace filament::backend {
 
 class VulkanPlatform;
@@ -169,6 +165,15 @@ private:
     math::mat3f getStreamTransformMatrix(Handle<HwStream> sh);
 
 
+    // This maps a VulkanSwapchain to a native swapchain. VulkanSwapchain should have a copy of the
+    // Platform::Swapchain pointer, but queryFrameTimestamps() and queryCompositorTiming() are
+    // synchronous calls, making access to VulkanSwapchain unsafe (this difference vs other backends
+    // is due to the ref-counting of vulkan resources).
+    struct {
+        std::mutex lock;
+        std::unordered_map<HandleId, Platform::SwapChain*> nativeSwapchains;
+    } mTiming;
+
     // This is necessary for us to write to push constants after binding a pipeline.
     using DescriptorSetLayoutHandleList = std::array<resource_ptr<VulkanDescriptorSetLayout>,
             VulkanDescriptorSetLayout::UNIQUE_DESCRIPTOR_SET_COUNT>;
@@ -204,13 +209,6 @@ private:
     bool const mIsSRGBSwapChainSupported;
     bool const mIsMSAASwapChainSupported;
     backend::StereoscopicType const mStereoscopicType;
-
-    // setAcquiredImage is a DECL_DRIVER_API_SYNCHRONOUS_N which means we don't necessarily have the
-    // data to process it at call time. So we store it and process it during updateStreams.
-    std::vector<resource_ptr<VulkanStream>> mStreamsWithPendingAcquiredImage;
-#if defined(__ANDROID__)
-    AndroidProducerThrottling mProducerThrottling;
-#endif
 };
 
 } // namespace filament::backend