+++ tools/matc/src/matc/MaterialCompiler.cpp
@@ -16,8 +16,6 @@
 
 #include "MaterialCompiler.h"
 
-#include "DirIncluder.h"
-
 #include <memory>
 #include <iostream>
 #include <utility>
@@ -52,8 +50,6 @@ bool MaterialCompiler::run(const matp::Config& config) {
     }
     std::unique_ptr<const char[]> buffer = input->read();
 
-    mParser.processTemplateSubstitutions(config, size, buffer);
-
     utils::Path const materialFilePath = utils::Path(input->getName()).getAbsolutePath();
     assert(materialFilePath.isFile());
 
@@ -69,16 +65,32 @@ bool MaterialCompiler::run(const matp::Config& config) {
         glslang::FinalizeProcess();
         return success;
     }
+    auto [resolvedStatus, resolvedString] =
+            mParser.resolveIncludes(buffer, size, materialFilePath,
+            config.getInsertLineDirectives(),
+            config.getInsertLineDirectiveChecks());
+
+    // It has failed to resolve the include directives.
+    if (!resolvedStatus.isOk()) {
+        std::cerr << resolvedStatus.getMessage() << std::endl;
+        return false;
+    }
+
+    // Now that the buffer is mutated, we need to update the buffer pointer and the size
+    // before parsing.
+    size = resolvedString.size();
+    auto modifiedBuffer = std::make_unique<char[]>(size);
+    std::strncpy(modifiedBuffer.get(), resolvedString.c_str(), size);
+    buffer = std::move(modifiedBuffer);
+
+    if (config.getOutputFormat() == matp::Config::OutputFormat::MAT) {
+        return writeMat(buffer, size, config);
+    }
+
+    mParser.processTemplateSubstitutions(config, size, buffer);
 
     MaterialBuilder::init();
     MaterialBuilder builder;
-
-    // Set the root include directory to the directory containing the material file.
-    DirIncluder includer;
-    includer.setIncludeDirectory(materialFilePath.getParent());
-
-    builder.includeCallback(includer)
-            .fileName(materialFilePath.getName().c_str());
 
     utils::Status status = mParser.parse(builder, config, size, buffer);
     if (!status.isOk()) {
@@ -133,7 +145,6 @@ bool MaterialCompiler::checkParameters(const matp::Config& config) {
     return true;
 }
 
-// this can either be here or material parser
 bool MaterialCompiler::compileRawShader(const char* glsl, size_t size, bool isDebug,
         matp::Config::Output* output, const char* ext) const noexcept {
     using namespace glslang;